wifi and ota working

+++ src/modes/HoneyOrangeMode.cpp
@@ -28,7 +28,7 @@ extern const TProgmemRGBPalette16 HoneyOrange_p FL_PROGMEM =
     HONEY_YELLOW,
 };
 
-CHoneyOrangeMode::CHoneyOrangeMode(const uint8_t numLeds, const String name)
+CHoneyOrangeMode::CHoneyOrangeMode(const uint16_t numLeds, const String name)
 : CPaletteMode(numLeds, name, HoneyOrange_p, 1) {
     
 }
