+++ src/modes/PaletteMode.cpp
@@ -1,14 +1,14 @@
 #include "PaletteMode.h"
 
-CPaletteMode::CPaletteMode(const uint8_t numLeds, const String name, const TProgmemRGBPalette16& palette, const uint8_t increment)
+CPaletteMode::CPaletteMode(const uint16_t numLeds, const String name, const TProgmemRGBPalette16& palette, const float increment)
 : CBaseMode(numLeds, name), increment(increment), palette(palette), blendType(LINEARBLEND), delay(15) {
 }
 
-CPaletteMode::CPaletteMode(const uint8_t numLeds, const String name, const TProgmemRGBPalette16& palette, const uint8_t increment, const TBlendType blendType)
+CPaletteMode::CPaletteMode(const uint16_t numLeds, const String name, const TProgmemRGBPalette16& palette, const float increment, const TBlendType blendType)
 : CBaseMode(numLeds, name), increment(increment), palette(palette), blendType(blendType), delay(15) {
 }
 
-CPaletteMode::CPaletteMode(const uint8_t numLeds, const String name, const TProgmemRGBPalette16& palette, const uint8_t increment, const TBlendType blendType, const unsigned long delay)
+CPaletteMode::CPaletteMode(const uint16_t numLeds, const String name, const TProgmemRGBPalette16& palette, const float increment, const TBlendType blendType, const unsigned long delay)
 : CBaseMode(numLeds, name), increment(increment), palette(palette), blendType(blendType), delay(delay) {
 }
 
@@ -19,9 +19,9 @@ void CPaletteMode::draw(CRGB *leds) {
         startIndex = startIndex + 1;
     }
 
-    uint8_t ci = startIndex;
-    for( int i = 0; i < numLeds; i++) {
-        leds[i] = ColorFromPalette( palette, ci, 255, blendType);
+    float ci = startIndex;
+    for( uint16_t i = 0; i < numLeds; i++) {
+        leds[i] = ColorFromPalette( palette, (uint8_t)ci, 255, blendType);
         ci+=increment;
     }
 }
