+++ src/modes/PaletteMode.h
@@ -6,14 +6,14 @@ class CPaletteMode : public CBaseMode {
 
 private:
     uint8_t startIndex = 0;
-    const uint8_t increment;
+    const float increment;
     const TProgmemRGBPalette16& palette;
     const TBlendType blendType;
     const unsigned long delay;
 
 public:
-	CPaletteMode(const uint8_t numLeds, const String name, const TProgmemRGBPalette16& palette, const uint8_t increment);
-    CPaletteMode(const uint8_t numLeds, const String name, const TProgmemRGBPalette16& palette, const uint8_t increment, const TBlendType blendType);
-    CPaletteMode(const uint8_t numLeds, const String name, const TProgmemRGBPalette16& palette, const uint8_t increment, const TBlendType blendType, const unsigned long delay);
+	CPaletteMode(const uint16_t numLeds, const String name, const TProgmemRGBPalette16& palette, const float increment);
+    CPaletteMode(const uint16_t numLeds, const String name, const TProgmemRGBPalette16& palette, const float increment, const TBlendType blendType);
+    CPaletteMode(const uint16_t numLeds, const String name, const TProgmemRGBPalette16& palette, const float increment, const TBlendType blendType, const unsigned long delay);
     virtual void draw(CRGB *leds);
 };
